Reminder of changes required by the Windows port. This solves the issues 228, 229, 230 and 232.

+++ src/extension/default_eventloop.c
@@ -25,6 +25,8 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
+#include "Config.h"
+
 #include "extension/default_eventloop.h"
 #include "debug.h"
 
@@ -32,7 +34,7 @@ static uint64_t get_now_plus(uint64_t amount)
 {
 	struct timeval tv;
 	uint64_t       now;
-	
+
 	if (gettimeofday(&tv, NULL)) {
 		perror("gettimeofday() failed");
 		exit(EXIT_FAILURE);
@@ -81,8 +83,7 @@ default_eventloop_schedule(getdns_eventloop *loop,
 #endif
 		default_loop->fd_events[fd] = event;
 		default_loop->fd_timeout_times[fd] = get_now_plus(timeout);
-		event->ev = (void *) (intptr_t) fd + 1;
-
+		event->ev = (void *)(intptr_t)(fd + 1);
 		DEBUG_SCHED( "scheduled read/write at %d\n", fd);
 		return GETDNS_RETURN_GOOD;
 	}
@@ -101,9 +102,8 @@ default_eventloop_schedule(getdns_eventloop *loop,
 	for (i = 0; i < MAX_TIMEOUTS; i++) {
 		if (default_loop->timeout_events[i] == NULL) {
 			default_loop->timeout_events[i] = event;
-			default_loop->timeout_times[i] = get_now_plus(timeout);
-			event->ev = (void *) (intptr_t) i + 1;
-
+			default_loop->timeout_times[i] = get_now_plus(timeout);		
+			event->ev = (void *)(intptr_t)(i + 1);
 			DEBUG_SCHED( "scheduled timeout at %d\n", (int)i);
 			return GETDNS_RETURN_GOOD;
 		}
@@ -219,8 +219,8 @@ default_eventloop_run_once(getdns_eventloop *loop, int blocking)
 		tv.tv_sec = 0;
 		tv.tv_usec = 0;
 	} else {
-		tv.tv_sec  = (timeout - now) / 1000000;
-		tv.tv_usec = (timeout - now) % 1000000;
+		tv.tv_sec  = (long)((timeout - now) / 1000000);
+		tv.tv_usec = (long)((timeout - now) % 1000000);
 	}
 	if (select(max_fd + 1, &readfds, &writefds, NULL,
 	    (timeout == ((uint64_t)-1) ? NULL : &tv)) < 0) {
